fixed agent class and commented examaple code

+++ include/config.hpp
@@ -2,15 +2,15 @@
 const float DELTA_TIME = 1.0f;
 
 /* WINDOW */
-const unsigned int WINDOW_WIDTH = 800;
-const unsigned int WINDOW_HEIGHT = 800;
+const unsigned short WINDOW_WIDTH = 800;
+const unsigned short WINDOW_HEIGHT = 800;
 
 /* AGENT */
 const unsigned int AGENT_COUNT = 100;
 const float AGENT_SPEED = 1.0f;
 
 /* AGENT - SENSOR */
-const unsigned int SENSOR_SIZE = 1;
+const unsigned short SENSOR_SIZE = 1;
 const float SENSOR_DISTANCE = 3.0f;
 const float SENSOR_ANGLE = 1.0f;
